Fixing Chaincode errors
