+++ fabric-chrome-extension/src/popup.html
@@ -14,7 +14,7 @@
 		  
 		<p>Adding Peers:</p>
 		<ul class="peer_list">      
-		  </ul>
+		</ul>
 		  
 		</div>
